codeAreaをコンポーネント化
